Move gesture classes to separate files Rename SimpleGesture to BaseGesture
