+++ src/handlers/gestures/rotationGesture.ts
@@ -1,0 +1,20 @@
+import { UnwrappedGestureHandlerEvent } from '../gestureHandlerCommon';
+import { BaseGesture } from './baseGesture';
+import { RotationGestureHandlerEventPayload } from '../RotationGestureHandler';
+
+export class RotationGesture extends BaseGesture<RotationGestureHandlerEventPayload> {
+  constructor() {
+    super();
+
+    this.handlerName = 'RotationGestureHandler';
+  }
+
+  setOnUpdate(
+    callback: (
+      event: UnwrappedGestureHandlerEvent<RotationGestureHandlerEventPayload>
+    ) => void
+  ) {
+    this.handlers.onUpdate = callback;
+    return this;
+  }
+}
